Rename ExpandedLearningResourceDisplay to LearningResourceExpanded
