+++ frontends/mit-open/src/page-components/LearningResourceDrawer/LearningResourceDrawer.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react"
-import { RoutedDrawer, ExpandedLearningResourceDisplay } from "ol-components"
+import { RoutedDrawer, LearningResourceExpanded } from "ol-components"
 import type { RoutedDrawerProps } from "ol-components"
 import { useLearningResourcesDetail } from "api/hooks/learningResources"
 import { imgConfigs } from "@/common/constants"
@@ -43,7 +43,7 @@ const DrawerContent: React.FC<{
   useCapturePageView(Number(resourceId))
 
   return (
-    <ExpandedLearningResourceDisplay
+    <LearningResourceExpanded
       imgConfig={imgConfigs.large}
       resource={resource.data}
     />
